backend/requirements.txt
Django>=3.0.10
psycopg2-binary>=2.8
djangorestframework>=3.11.1
djangorestframework-camel-case>=1.2.0
django-filter>=2.4.0